requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
coverage==5.5
deprecation==2.1.0
Django==5.0.7
eyed3==0.9.7
filetype==1.2.0
h11==0.14.0
idna==3.7
lyricsgenius==3.0.1
mutagen==1.47.0
outcome==1.3.0.post0
packaging==24.1
pycryptodomex==3.20.0
PySocks==1.7.1
python-dotenv==1.0.1
redis==5.0.7
requests==2.32.3
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
toml==0.10.2
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
websockets==12.0
wget==3.2
wsproto==1.2.0
yt-dlp==2024.7.16
